feat: Add a salad recipe webpage to the website
Create a new html file for the salad recipe and add a salad image to the
images directory.

Populate the salad page with the recipe and salad image, keeping the
same general structure as the previous recipe pages.

Modify the homepage with a relative link to the new recipe page and add
all current links to an unordered list.